requirements.txt
pygame==2.0.0.dev6
PyOpenGL==3.1.0
